backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langfuse>=3.10.7",
    "langgraph>=1.0.4",
    "load-dotenv>=0.1.0",
    "pandas>=2.3.3",
    "requests>=2.31.0",
]
